The branch was 663 commits behind develop, which moved the package from
supervision/ to src/supervision/. Git did not detect that rename, so both
files this branch touches came through as modify/delete conflicts and would
otherwise have been resurrected in the retired layout.

Resolution:
- Removed supervision/detection/core.py and supervision/detection/vlm.py;
  their successors live under src/supervision/detection/.
- Took develop's tests/detection/test_vlm.py.

The Kosmos-2 feature itself is re-applied on top of the current VLM
architecture in the following commits.

---

+++ examples/time_in_zone/utils/timers.py
@@ -6,22 +6,22 @@ import supervision as sv
 
 
 class FPSBasedTimer:
-    """
-    A timer that calculates the duration each object has been detected based on frames
-    per second (FPS).
+    """A timer that calculates the duration each object has been detected based on
+    frames per second (FPS).
 
     Attributes:
-        fps (int): The frame rate of the video stream, used to calculate time durations.
+        fps (float): The frame rate of the video stream, used to calculate
+            time durations.
         frame_id (int): The current frame number in the sequence.
         tracker_id2frame_id (Dict[int, int]): Maps each tracker's ID to the frame number
             at which it was first detected.
     """
 
-    def __init__(self, fps: int = 30) -> None:
+    def __init__(self, fps: float = 30) -> None:
         """Initializes the FPSBasedTimer with the specified frames per second rate.
 
         Args:
-            fps (int): The frame rate of the video stream. Defaults to 30.
+            fps (float): The frame rate of the video stream. Defaults to 30.
         """
         self.fps = fps
         self.frame_id = 0
@@ -51,8 +51,7 @@ class FPSBasedTimer:
 
 
 class ClockBasedTimer:
-    """
-    A timer that calculates the duration each object has been detected based on the
+    """A timer that calculates the duration each object has been detected based on the
     system clock.
 
     Attributes:
